Move seo url related migrations from storefront to core
